pyproject.toml
[project]
name = "resklogits"
version = "0.1.0"
description = "GPU-accelerated shadow ban logits processor for LLM safety filtering"
authors = [
    {name = "RESK Team"}
]
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
keywords = ["llm", "safety", "filtering", "aho-corasick", "gpu", "logits", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "accelerate>=0.24.0",
    "types-PyYAML>=6.0",
]
accelerate = [
    "accelerate>=0.24.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
resklogits = "resklogits.cli:main"

[project.urls]
Homepage = "https://github.com/resk-team/resklogits"
Documentation = "https://github.com/resk-team/resklogits#readme"
Repository = "https://github.com/resk-team/resklogits"
Issues = "https://github.com/resk-team/resklogits/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resklogits"]
include = [
    "/src/resklogits/**/*.py",
    "/src/resklogits/**/*.json",
    "/src/resklogits/py.typed",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

